[ALLUXIO-1910] Move Format to alluxio.cli package
