+++ integration/bin/alluxio-worker-yarn.sh
@@ -18,7 +18,7 @@ echo "Formatting Alluxio Worker"
   -Dalluxio.home="${ALLUXIO_HOME}" \
   -Dalluxio.logger.type="WORKER_LOGGER" \
   -Dalluxio.logs.dir="${YARN_LOG_DIR}" \
-  alluxio.Format WORKER > "${YARN_LOG_DIR}"/worker.out 2>&1
+  alluxio.cli.Format WORKER > "${YARN_LOG_DIR}"/worker.out 2>&1
 
 echo "Starting Alluxio Worker"
