+++ integration/mesos/src/main/java/alluxio/mesos/AlluxioWorkerExecutor.java
@@ -11,7 +11,7 @@
 
 package alluxio.mesos;
 
-import alluxio.Format;
+import alluxio.cli.Format;
 import alluxio.underfs.UnderFileSystemRegistry;
 import alluxio.worker.AlluxioWorker;
